Add new feature utilizing pre-allocated moves buffer
- Pass in a pre-allocated moves buffer. This buffer is only used
inside the gen method and overwrites previous data with new move
data. Instead of passing this moves buffer back for use, allocate
a new array of exactly the size of the number of newly generated
moves in the moves buffer. This saves time always allocating a max
size array every gen method call and makes the usage of the
moves array more consistent where it needs to be used. Make sure
in multiple threads that each thread uses their own copy of a moves
buffer so there are no concurrency conflicts